gcs-celery-worker-src/requirements.txt
amqp==5.0.6
autopep8==1.5.7
billiard==3.6.4.0
cached-property==1.5.2
cachetools==4.2.2
celery==5.1.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
google-api-core==1.28.0
google-auth==1.30.1
google-cloud==0.34.0
google-cloud-core==1.6.0
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
idna==2.10
importlib-metadata==4.4.0
kombu==5.1.0
packaging==20.9
prompt-toolkit==3.0.18
protobuf==3.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyparsing==2.4.7
pytz==2021.1
redis==3.5.3
requests==2.25.1
rsa==4.7.2
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1
